+++ broker/src/main.rs
@@ -4,12 +4,13 @@ use actix_cors::Cors;
 use actix_web::middleware::from_fn;
 use actix_web::{get, web, App, HttpResponse, HttpServer};
 use api::{
-    add_pod_details, add_pods_batch, add_pods_syscalls, add_svc_details, establish_connection,
-    get_audit_verdicts, get_pod_by_ip, get_pod_by_name, get_pod_details, get_pod_syscall_name,
-    get_pod_traffic, get_pod_traffic_name, get_pods_by_node, get_seccomp_profile,
-    get_seccomp_profile_file, get_svc_by_ip, get_svc_details, get_version, list_seccomp_profiles,
-    mark_pod_dead, post_seccomp_node_status, set_statement_timeout, spawn_retention,
-    spawn_version_check, AuditClient, StatementTimeoutCustomizer, VersionCheckState,
+    add_node_facts, add_pod_details, add_pods_batch, add_pods_syscalls, add_svc_details,
+    establish_connection, get_audit_verdicts, get_pod_by_ip, get_pod_by_name, get_pod_details,
+    get_pod_syscall_name, get_pod_traffic, get_pod_traffic_name, get_pods_by_node,
+    get_seccomp_profile, get_seccomp_profile_file, get_svc_by_ip, get_svc_details, get_version,
+    list_seccomp_profiles, mark_pod_dead, post_seccomp_node_status, set_statement_timeout,
+    spawn_retention, spawn_version_check, AuditClient, StatementTimeoutCustomizer,
+    VersionCheckState,
 };
 
 use diesel::r2d2;
@@ -365,6 +366,7 @@ async fn main() -> Result<(), std::io::Error> {
             .service(get_pods_by_node)
             .service(get_audit_verdicts)
             .service(mark_pod_dead)
+            .service(add_node_facts)
             .service(get_version)
             .service(health_check)
             .service(metrics)